praise {praise}	R Documentation

Praise Users

Description

Build friendly R packages that praise their users if they have done something good, or they just need it to feel better.

Randomized praise based on a template

Usage

praise(template = "You are ${adjective}!")

Arguments

template
Character scalar, the template string.

Details

Replace parts of the template with random words from the praise word lists. See examples below.

Examples

praise()

## Capitalization
praise("${Exclamation}! This ${rpackage} is ${adjective}!")

## All upper case
praise("${EXCLAMATION}! You have done this ${adverb_manner}!")

[Package praise version 1.0.0 Index]